refactor: absorb more meteor declarations
