--- apps/meteor/server/lib/2fa/definitions/MeteorUser.d.ts
@@ -1,12 +1,0 @@
-declare module 'meteor/meteor' {
-	namespace Meteor {
-		interface UserServices {
-			totp?: {
-				enabled: boolean;
-				hashedBackup: string[];
-				secret: string;
-				tempSecret?: string;
-			};
-		}
-	}
-}
